Move all element classes to separate files
