+++ skfem/element/element_hex/__init__.py
@@ -1,0 +1,1 @@
+from .element_hex1 import ElementHex1
