+++ skfem/element/element_line/__init__.py
@@ -1,0 +1,1 @@
+from .element_line_p1 import ElementLineP1
