+++ skfem/element/element_quad/__init__.py
@@ -1,0 +1,2 @@
+from .element_quad1 import ElementQuad1
+from .element_quad2 import ElementQuad2
