+++ skfem/element/element_tet/__init__.py
@@ -1,0 +1,5 @@
+from .element_tet_p0 import ElementTetP0
+from .element_tet_p1 import ElementTetP1
+from .element_tet_p2 import ElementTetP2
+from .element_tet_rt0 import ElementTetRT0
+from .element_tet_n0 import ElementTetN0
