+++ skfem/element/element_tri/__init__.py
@@ -1,0 +1,7 @@
+from .element_tri_p1 import ElementTriP1
+from .element_tri_p2 import ElementTriP2
+from .element_tri_dg import ElementTriDG
+from .element_tri_p0 import ElementTriP0
+from .element_tri_rt0 import ElementTriRT0
+from .element_tri_morley import ElementTriMorley
+from .element_tri_argyris import ElementTriArgyris
