+++ skfem/element/element_tri/element_tri_p0.py
@@ -1,0 +1,11 @@
+from ..element_h1 import ElementH1
+
+
+class ElementTriP0(ElementH1):
+    interior_dofs = 1
+    dim = 2
+    maxdeg = 0
+    dofnames = ['u']
+
+    def lbasis(self, X, i):
+        return 1 + 0*X[0, :], 0*X
